requirements.txt
discord-py
hdate
markdownify
aiosqlite
geopy
tzwhere
bs4